+++ src/backend/joanie/tests/core/test_api_base.py
@@ -15,12 +15,12 @@ class BaseAPITestTestCase(BaseAPITestCase):
         the token attributes should correspond to the data of the user
         """
         user = factories.UserFactory(
-            username="Sam", email="sam@fun-test.fr", language="fr"
+            username="Sam", email="sam@fun-test.fr", language="fr-fr"
         )
         token = self.generate_token_from_user(user)
         self.assertEqual("sam@fun-test.fr", token.payload.get("email"))
         self.assertEqual("Sam", token.payload.get("username"))
-        self.assertEqual("fr", token.payload.get("language"))
+        self.assertEqual("fr-fr", token.payload.get("language"))
 
         # expiration date is over a day from now
         self.assertGreater(