+++ src/backend/joanie/tests/core/test_models_user.py
@@ -1,6 +1,7 @@
 """Test suite for badge models."""
-from django.db import IntegrityError
+from django.core.exceptions import ValidationError
 from django.test import RequestFactory, TestCase
+from django.test.utils import override_settings
 
 from joanie.core.factories import UserFactory
 from joanie.core.models import User
@@ -12,7 +13,7 @@ class UserModelTestCase(TestCase):
     def test_model_create(self):
         """A simple test to check model consistency."""
 
-        UserFactory(username="Sam", email="sam@fun-test.fr", language="fr")
+        UserFactory(username="Sam", email="sam@fun-test.fr", language="fr-fr")
         UserFactory(username="Joanie")
 
         self.assertEqual(User.objects.count(), 2)
@@ -20,8 +21,30 @@ class UserModelTestCase(TestCase):
         first_user = User.objects.first()
         self.assertEqual(first_user.username, "Sam")
         self.assertEqual(first_user.email, "sam@fun-test.fr")
-        self.assertEqual(first_user.language, "fr")
+        self.assertEqual(first_user.language, "fr-fr")
         self.assertEqual(str(first_user), "Sam")
+
+    @override_settings(
+        LANGUAGES=(("fr-ca", "Canadian"), ("it", "Italian")),
+        LANGUAGE_CODE="fr-ca",
+    )
+    def test_model_create_language(self):
+        """Check language is part of the languages
+        declared in the settings."""
+
+        UserFactory(username="Sam", email="sam@fun-test.fr", language="fr-ca")
+        UserFactory(username="Sam2", email="sam@fun-test.fr", language="it")
+
+        with self.assertRaises(ValidationError) as context:
+            UserFactory(username="Sam3", email="sam@fun-test.fr", language="fr-fr")
+
+        self.assertEqual(
+            "La valeur « 'fr-fr' » n’est pas un choix valide.",
+            str(context.exception.messages[0]),
+        )
+
+        with self.assertRaises(ValidationError):
+            UserFactory(username="Sam", email="sam@fun-test.fr", language="en-us")
 
     def test_models_unique_username(self):
         """
@@ -29,8 +52,12 @@ class UserModelTestCase(TestCase):
         """
         UserFactory(username="Sam", email="sam@fun-test.fr")
 
-        with self.assertRaises(IntegrityError):
+        with self.assertRaises(ValidationError) as error:
             UserFactory(username="Sam")
+            self.assertEqual(
+                "{'username': ['A user with that username already exists.']}",
+                str(error.exception),
+            )
 
     def test_models_multiple_emails(self):
         """
@@ -57,7 +84,7 @@ class UserModelTestCase(TestCase):
 
         # email has been updated
         self.assertEqual(user.email, "sam@fun-test.fr")
-        self.assertEqual(user.language, "fr")
+        self.assertEqual(user.language, "fr-fr")
         # no new object has been created
         self.assertEqual(User.objects.count(), 1)
 
@@ -67,5 +94,80 @@ class UserModelTestCase(TestCase):
         # a new object has been created
         self.assertEqual(User.objects.count(), 2)
         user2 = User.objects.get(username="Sam2")
-        self.assertEqual(user2.language, "en")
+        self.assertEqual(user2.language, "en-us")
         self.assertEqual(user2.email, "sam@fun-test.fr")
+
+    def test_models_create_or_update_from_request_language(self):
+        """
+        Check using the method create_or_update_from_request, the language
+        is set depending of the values of the language available in the
+        settings and understood in different ways
+        """
+        request = RequestFactory()
+        request.username = "Sam"
+        request.email = "sam@fun-test.fr"
+
+        # with _
+        request.language = "fr_FR"
+        User.update_or_create_from_request_user(request)
+        user = User.objects.get(username="Sam")
+        user.refresh_from_db()
+        self.assertEqual(user.language, "fr-fr")
+
+        request.language = "en"
+        User.update_or_create_from_request_user(request)
+        user.refresh_from_db()
+        self.assertEqual(user.language, "en-us")
+
+        request.language = "en-gb"
+        User.update_or_create_from_request_user(request)
+        user.refresh_from_db()
+        self.assertEqual(user.language, "en-us")
+
+        request.language = "fr"
+        User.update_or_create_from_request_user(request)
+        user.refresh_from_db()
+        self.assertEqual(user.language, "fr-fr")
+
+        # `it` is not defined in the settings
+        request.language = "it"
+        User.update_or_create_from_request_user(request)
+        user.refresh_from_db()
+        # the default language is used
+        self.assertEqual(user.language, "en-us")
+
+        request.language = "whatever"
+        User.update_or_create_from_request_user(request)
+        user.refresh_from_db()
+        # the default language is used
+        self.assertEqual(user.language, "en-us")
+
+        with override_settings(
+            LANGUAGES=(("fr-ca", "Canadian"), ("it", "Italian"), ("es-ve", "Spain")),
+            LANGUAGE_CODE="es-ve",
+        ):
+            request.language = "fr_FR"
+            User.update_or_create_from_request_user(request)
+            user.refresh_from_db()
+            self.assertEqual(user.language, "fr-ca")
+
+            request.language = "fr-FR"
+            User.update_or_create_from_request_user(request)
+            user.refresh_from_db()
+            self.assertEqual(user.language, "fr-ca")
+
+            request.language = "fr"
+            User.update_or_create_from_request_user(request)
+            user.refresh_from_db()
+            self.assertEqual(user.language, "fr-ca")
+
+            request.language = "it"
+            User.update_or_create_from_request_user(request)
+            user.refresh_from_db()
+            self.assertEqual(user.language, "it")
+
+            request.language = "ru"
+            User.update_or_create_from_request_user(request)
+            user.refresh_from_db()
+            # default is used
+            self.assertEqual(user.language, "es-ve")