+++ CHANGELOG.md
@@ -58,6 +58,7 @@ and this project adheres to
 
 ### Changed
 
+- Configure language field of the user's model
 - Update codebase to remove the use of deprecated pytz and USE_L10N
   with Django 4.0
 - Refactor links between enrollment and order models